+++ meta_package_manager/bar_plugin_renderer.py
@@ -41,8 +41,6 @@ from unittest.mock import patch
 from boltons.iterutils import flatten
 from click_extra import echo
 from click_extra.table import TableFormat, render_table
-from click_extra.theme import get_current_theme as theme
-
 from .bar_plugin import MPMPlugin
 from .pool import pool
 
@@ -233,8 +231,8 @@ class BarPluginRenderer(MPMPlugin):
                     # Fallback on mpm itself which is capable of simulating a full
                     # upgrade.
                     logging.info(
-                        f"{theme().invoked_command(manager_id)} "
-                        "does not implement upgrade_all_cli.",
+                        "Does not implement upgrade_all_cli.",
+                        extra={"label": manager_id},
                     )
                     mpm_args, _runnable, _up_to_date, _version, _error = self.best_mpm
                     upgrade_all_cli = (